Remove :metric_f from user-facing event representation
Before this change, the event maps the user would be working with contained both
:metric and :metric_f.

This is confusing when using something like (default ...) because it's not clear
which one you need to set (or even if you need to set *both*).

:metric_f serves no real purpose at this point, because it's overwritten with
the value of :metric in pre-dump-event anyway.

This change removes :metric_f from the user-facing events, so users always work
with :metric in their riemann.config files.  The fact that "metric_f" is a thing
becomes an implementation detail.