Basic websockets pubsub server.
You can set up a websocket server with

(ws-server)

or

(ws-server :host x :port 5556)

The server answers websocket connections to /pubsub/<topic>. Any query
can be applied as a filter:

/pubsub/topic/query=true /pubsub/topic/query=(service =~ "%cat" and
metric < 5.4) (query must be url-encoded, obviously)

The resulting websocket will receive a stream of UTF8 frames, each
containing a JSON representation of an event from that topic which
matched the query.

Initial testing shows excellent latency (no perceivable lag) but appears
to drop events in volume. Need to investigate queue depths. Be aware
that the correctness and safety of the query system are not guaranteed;
just like the protobufs server, assume this system runs *arbitrary* code
from the network.