Critical bugfix: (where) evaluated its children every time.
When I added (else) blocks to (where) in 0.1.3, I inadvertently
introduced a bug which caused children of (where) to be evaluated
*every* time. This breaks all stateful streams like ddt, rate, etc. It's
a straightforward fix. I've also added a test explicitly confirming
that (by) evaluates its children at the correct times.

Future changes to macros will be accompanied by careful
time-of-evaluation tests. This stuff is surprisingly tricky to do right.