Switch to the builtin Netty protobuf codec.
18% performance improvement in single-threaded TCP round-trip time.
EIGHTEEN PERCENT! YEeeeEEeEeAAaaAaAaaaAAaAAAHHhHHH!

We used to use clj-protobuf, back in the day, which meant we had to
decode netty buffers into protobuf objects, and write them back to
bytebuffers for the return trip. That required extra copies when we
could read/write directly from the Netty buffers. Since I switched out
clj-protobuf for the regular google protobuf classes, and worked out all
the compilation/namespace issues around that, we're free to get rid of
that parsing layer and switch to the built-in Netty protobuf codec.