(streams/top): streaming top-k
(streams/top 2 :metric index prn) indexes the top 2 host/service pairs
by metric, and prints events which are *not* in the top 2. Closes #119.

Also introduces (common/expire), which returns an expired copy of an
event, and (streams/dual pred true-stream bottom-stream), which
maintains a pair of streams, one for (pred e) and one for (not (pred
e)), making sure that both have a consistent view of the world.