PUT /events: submits JSON events to Riemann.
(ws-server) learned a new trick! You can now issue an HTTP PUT to
Riemann with a list of (optionally whitespace-separated) JSON objects
which will be converted to events and applied to streams. For example:

{"service": "bar"}
{"service": "foo"
 "time": "2013-04-15T18:06:58-07:00"
 "metric": 1.23
 "cat": "mew!"}
```

curl -X PUT -v -T events.json http://localhost:5556/events

The response is a series of JSON maps in the same format: empty if the
event was accepted successfully, or an error like {"error": "something
went wrong"}.

This API is fully streaming: you are encouraged to hold the connection
body open and send events continuously. It'll push somewhere around
30-60K events/sec over a connection depending on size.