Remove openjdk dependencies for more flexibility
This came up when using the opscode-cookbooks/java cookbook. If you install the oracle flavor, which installs via tarballs, the riemann deb will install but will break apt-get for everything else. It won't let you install anything else because riemann doesn't have openjdk like it expects. 

I vote that we remove java as a strict dependency to sidestep this problem.