Add support for vector of events in Librato
The Librato gauge and counter stream have been modified to accept a vectorof
events. This provides the capability to send multiple metrics in a single
http POST.

A test for multiple events in a single :gauge call has been added.
An analgous set of tests for :counter have been added.