Provide a websocket transport separated from aleph
This is the first of two steps meant to get rid of the dependency
on aleph and lamina in standard riemann (mostly to provide a more
lightweight uberjar and much faster startup times).

This implements the websocket functionality on top of httpkit which
will also provide a good foundation to provide the SSE functionality.

The lack of feedback from httpkit prevents out latency statistics
from being generated.