Upgrade to netty 4.0.21
Open Tasks:

- Decide what to do with shared-event-executor (src/riemann/transport.clj).

The shared-event-executor is replacing the old thread pool (because it was
deprecated).
But there might be a better implementation available which gives you a similar
degree of control.

This should probably be addressed before the next release of riemann.

The current value is set to 100 (100 available threads) using the
DefaultEventExecutorGroup.

- Test cases might be misbehaving.

riemann.graphite-test seems to be displaying an error which might be
a programming error. An exception that is not being caught by implemention the
correct pipeline event handler.

I'm not sure how to address this.