Also run the test suite on PR
We have CI running when code is pushed to the riemann repository.  For
external contribution, this only happen when a PR is merged, but
detecting issues earlier is advisable.

Instead of running CI when code is pushed in any branch of the repo, run
CI when the main branch is updated, and when a PR targeting the main
branch is updated.

Signed-off-by: Romain Tartière <romain@blogreen.org>